+++ libraries/AP_Math/AP_Math.h
@@ -9,7 +9,7 @@
 #include <AP_Param/AP_Param.h>
 
 #include "definitions.h"
-#include "edc.h"
+#include "crc.h"
 #include "matrix3.h"
 #include "polygon.h"
 #include "quaternion.h"
