+++ libraries/AP_Math/crc.h
@@ -22,3 +22,7 @@ uint8_t crc_crc8(const uint8_t *p, uint8_t len);
 uint16_t crc_xmodem_update(uint16_t crc, uint8_t data);
 uint16_t crc_xmodem(const uint8_t *data, uint16_t len);
 uint32_t crc_crc32(uint32_t crc, const uint8_t *buf, uint32_t size);
+
+// Copyright (C) 2010 Swift Navigation Inc.
+// Contact: Fergus Noble <fergus@swift-nav.com>
+uint16_t crc16_ccitt(const uint8_t *buf, uint32_t len, uint16_t crc);
